Modified this script to skip file concatenation (step 9 in the previous version)
On my workstation there was an issue where the output files (_all.fasta) contained the same set of sequences duplicated twice.

I believe this was resulting from a race condition where sometimes (or certain systems) the contents of this file would be written
before being read and then written out again.

In any case, just writing out the observed sequecnes and the ref sequences to the same file initially
and avoiding concatenation solved the problem.

This addresses issue #35.  Closes #35